- moved components from index to according dir - moved functions to utils dir - replaced package-lock.json by yarn.lock - added files to .gitignore
