+++ .gitignore
@@ -8,6 +8,12 @@
 # testing
 /coverage
 
+# idea
+.idea
+
+# eslintcache
+.eslintcache
+
 # production
 /build
