+++ src/components/Square.jsx
@@ -1,0 +1,20 @@
+import React from 'react'
+
+const Square = (props) => {
+  const {
+    onClick,
+    position,
+    value,
+  } = props
+
+  return (
+    <button
+      className={`square ${position}`}
+      onClick={onClick}
+    >
+      {value}
+    </button>
+  )
+}
+
+export default Square
