+++ src/utils/reportWebVitals.js
@@ -1,0 +1,18 @@
+import { getCLS, getFID, getFCP, getLCP, getTTFB } from 'web-vitals'
+
+const reportWebVitals = (onPerfEntry) => {
+  if (onPerfEntry && onPerfEntry instanceof Function) {
+    /** imports are always at top of file
+     * there is no case when you must do like you did
+     * also, here is an examples of that functions usage
+     *  https://www.npmjs.com/package/web-vitals
+     *  */
+    getCLS(onPerfEntry);
+    getFID(onPerfEntry);
+    getFCP(onPerfEntry);
+    getLCP(onPerfEntry);
+    getTTFB(onPerfEntry);
+  }
+};
+
+export default reportWebVitals;
